get rid of TreeGridCollection in favor of function in useTreeGridState
TreeGridCollection was only really being used to generate a keymap of tree grid nodes to set aria-posinset and other aria attributes so it was a bit overkill and didnt need all the extra required getters, especially since TableCollection would be used in layout/keyboard delegates.
