+++ packages/@react-stately/collections/src/index.ts
@@ -16,3 +16,4 @@ export {Section} from './Section';
 export {useCollection} from './useCollection';
 export {getItemCount} from './getItemCount';
 export {getChildNodes, getFirstItem, getLastItem, getNthItem, compareNodeOrder} from './getChildNodes';
+export {CollectionBuilder} from './CollectionBuilder';
